[#8] Move `enableSecureFlag` ext into RecentAppsThumbnailHidingActivity
